feat: Cria o pacote controller seguindo SRP e modularização do código.
